Start refactoring inventory
