+++ scene/rooms/Room.cpp
@@ -17,6 +17,10 @@ Room::Room(std::string name, std::shared_ptr<Movement> m, uint8_t mask, int64_t 
 Room::Room(std::string name, std::string description, std::shared_ptr<Movement> m, uint8_t mask, int64_t nice)
     : Menu(std::vector<std::string>()), roomName(std::move(name)), description(std::move(description)), nice(nice), movement(std::move(m)) {
 
+    // The first option is always accessing the inventory
+    entries.emplace_back("Inventory");
+    funcs.emplace_back([](){  });
+
     // A mask of 0bx0x0 means we're in the last room - proceeding is impossible
     // The only movement option should be next floor or save, there is no turning back
     if ((mask & 0b0101) == 0) {
